requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.65
botocore==1.34.65
celery==5.3.6
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-storages==1.14.2
djangorestframework==3.15.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.5
msgpack==1.0.8
orderedmultidict==1.0.1
phonenumbers==8.13.32
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
redis==5.0.3
s3transfer==0.10.1
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
